requirements.txt
argcomplete
gitpython
importlib_resources
npyscreen
pyyaml